refactor: move vlad code to its own module so that we can have a VLAD histogram cache
